project02_08_04
I started the second project project02_08_04. I built out the form for the interviewer and the candidate. I then created the three main functions needed to connect to ,select, and create the table. i made all the variables needed for the inputs and functions.I started to make the code to show the form display or not too.